Backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
httpx
numpy
scikit-learn
sentence-transformers
phonetics
nltk
sqlalchemy
databases[postgresql]
prometheus-client
structlog
python-json-logger
faiss-cpu
python-levenshtein
